modified bash config files and added newsboat files

+++ bash_logout
@@ -2,14 +2,17 @@
 
 # This file is sourced when a login shell terminates.
 
-if [ -n "$SSH_AUTH_SOCKET" ]
+if [ "$TERM" == "linux" ]
 then
-    eval "$(/usr/bin/ssh-agent -k)"
-fi
+    if [ -n "$SSH_AUTH_SOCKET" ]
+    then
+        eval "$(/usr/bin/ssh-agent -k)"
+    fi
 
-if [ -d "$XDG_RUNTIME_DIR" ]
-then
-    echo y | rm -r "$XDG_RUNTIME_DIR"
+    if [ -d "$XDG_RUNTIME_DIR" ]
+    then
+        echo y | rm -r "$XDG_RUNTIME_DIR"
+    fi
 fi
 
 # Clear the screen for security's sake.
